Don't show LTI button on locked submission
refs: MBL-13324
affects: Student
release note: Fixed an issue where LTI submissions could still be accessed when locked
