--- Student/Canvas/Pages/Page+ViewModel.swift
@@ -1,35 +1,0 @@
-//
-// This file is part of Canvas.
-// Copyright (C) 2016-present  Instructure, Inc.
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as
-// published by the Free Software Foundation, either version 3 of the
-// License, or (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program.  If not, see <https://www.gnu.org/licenses/>.
-//
-
-import UIKit
-import CanvasCore
-import ReactiveSwift
-
-extension Page {
-    static func colorfulPageViewModel(session: Session, page: Page) -> ColorfulViewModel {
-        let vm = ColorfulViewModel(features: page.frontPage ? [.token] : [])
-        vm.title.value = page.title
-        if page.frontPage {
-            vm.tokenViewText.value = NSLocalizedString("Front Page", comment: "badge indicating front page")
-        }
-        vm.color <~ session.enrollmentsDataSource.color(for: page.contextID)
-
-        return vm
-    }
-}
-
